+++ src/EventSubscriber/CommentNotificationSubscriber.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace App\EventListener;
+namespace App\EventSubscriber;
 
 use App\Entity\Comment;
 use App\Events;
